requirements.txt
llama-cpp-python
huggingface-hub         
mysql-connector-python  
python-dotenv           
